Solving the dragmove() problem and creating a folder for the Graphical UI
