--- LibraryProject/LibraryApp/MainWindow.xaml
@@ -1,57 +1,0 @@
-﻿<Window x:Class="LibraryApp.MainWindow"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:materialDesign="http://materialdesigninxaml.net/winfx/xaml/themes"
-        xmlns:local="clr-namespace:LibraryApp"
-        mc:Ignorable="d"
-        Title="MainWindow" Height="450" Width="800" WindowStartupLocation="CenterScreen" WindowStyle="None"
-        ResizeMode="NoResize">
-    <Grid MouseDown="UIElement_OnMouseDown">
-        <Grid.ColumnDefinitions>
-            <ColumnDefinition Width="21*"/>
-            <ColumnDefinition Width="379*"/>
-        </Grid.ColumnDefinitions>
-        <Grid.Background>
-            <LinearGradientBrush StartPoint="0.1,0" EndPoint="0.9,1">
-                <GradientStop Color="#8bbdd9" Offset="1"/>
-                <GradientStop Color="#fff" Offset="0"/>
-            </LinearGradientBrush>
-        </Grid.Background>
-        <Border Style="{StaticResource StartupPagesStyle}"  Grid.ColumnSpan="2">
-        </Border>
-        <StackPanel Orientation="Horizontal" Grid.ColumnSpan="2" Grid.Column="0">
-            <StackPanel Width="350">
-                <Image Source="pack://application:,,,/Image/ReadingBook.png" Width="300" VerticalAlignment="Top" Margin="30" Stretch="Fill" OpacityMask="Black"/>
-            </StackPanel>
-            <StackPanel Width="350">
-                <StackPanel Margin="20 40">
-                    <TextBlock Text="All booked up!" Style="{StaticResource MainMessage}"/>
-                    <StackPanel Style="{StaticResource PrimaryStackPanel}">
-                        <Button Style="{StaticResource GoToPageBtn}"
-                               ToolTip="Login" 
-                               x:Name="BtnLogin"
-                               Click="BtnGoToLogin_OnClick" 
-                               Content="Login" />
-                    </StackPanel>
-                    <StackPanel Style="{StaticResource PrimaryStackPanel}">
-                        <Button Style="{StaticResource GoToPageBtn}"
-                                ToolTip="Register" 
-                                x:Name="BtnRegister" 
-                                Click="BtnRegister_OnClick" 
-                                Content="Register" ></Button>
-                    </StackPanel>
-                </StackPanel>
-            </StackPanel>
-            <StackPanel Width="100">
-                <Button Margin="10 10" Style="{StaticResource MaterialDesignFloatingActionButton}" Background="{x:Null}"
-                        ToolTip="Close" x:Name="BtnExit" Click="BtnExit_OnClick">
-                    <materialDesign:PackIcon Kind="Close" Foreground="AliceBlue" Width="30" Height="30"/>
-                </Button>
-            </StackPanel>
-        </StackPanel>
-        
-    </Grid>
-
-</Window>
