+++ LibraryProject/LibraryApp/View/AdminHome.xaml
@@ -1,4 +1,4 @@
-﻿<UserControl x:Class="LibraryApp.HomeAdmin"
+﻿<Window      x:Class="LibraryApp.View.AdminHome"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
@@ -6,7 +6,8 @@
              xmlns:materialDesign="http://materialdesigninxaml.net/winfx/xaml/themes"
              xmlns:local="clr-namespace:LibraryApp"
              mc:Ignorable="d" 
-             d:DesignHeight="800" d:DesignWidth="1200">
+             d:DesignHeight="800" d:DesignWidth="1200" WindowStartupLocation="CenterScreen" WindowStyle="None"
+             ResizeMode="NoResize">
     <Grid>
         <Grid.ColumnDefinitions>
             <ColumnDefinition Width="0.35*"/>
@@ -18,7 +19,7 @@
                 <RowDefinition Height="*"/>
             </Grid.RowDefinitions>
             <TextBlock Text="All Booked Up" Style="{StaticResource TitleStyle}"/>
-			<StackPanel Grid.Row="1">
+            <StackPanel Grid.Row="1">
                 <Button x:Name="AddBookBtn" Content="Add New Book" Style="{StaticResource MaterialDesignRaisedButton}" materialDesign:ButtonAssist.CornerRadius="10"
                         BorderThickness="2" BorderBrush="AliceBlue" Click="AddBookBtn_Click" Width="200"/>
 
@@ -28,10 +29,10 @@
                         <Button x:Name="DeleteCustomerBtn" Content="View books report" Style="{StaticResource SecondaryActionBtn}" />
                     </StackPanel>
                 </Expander>
-                </StackPanel>
+            </StackPanel>
         </Grid>
-		<Grid Grid.Column="1" Background="#e8f1f5">
-			<Grid.RowDefinitions>
+        <Grid Grid.Column="1" Background="#e8f1f5">
+            <Grid.RowDefinitions>
                 <RowDefinition Height="0.20*"/>
                 <RowDefinition Height="*"/>
             </Grid.RowDefinitions>
@@ -53,6 +54,6 @@
             </StackPanel>
 
         </Grid>
-        
-        </Grid>
-</UserControl>
+
+    </Grid>
+</Window>
