+++ LibraryProject/LibraryApp/View/RegisterUser.xaml
@@ -1,4 +1,4 @@
-﻿<UserControl x:Class="LibraryApp.Register"
+﻿<Window      x:Class="LibraryApp.View.RegisterUser"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
@@ -6,7 +6,8 @@
              xmlns:materialDesign="http://materialdesigninxaml.net/winfx/xaml/themes"
              xmlns:local="clr-namespace:LibraryApp"
              mc:Ignorable="d" 
-             d:DesignHeight="450" d:DesignWidth="800">
+             Title="Register" Height="450" Width="800" WindowStartupLocation="CenterScreen" WindowStyle="None"
+             ResizeMode="NoResize">
     <Grid MouseDown="UIElement_OnMouseDown">
         <Grid.ColumnDefinitions>
             <ColumnDefinition Width="Auto"/>
@@ -65,4 +66,4 @@
         </StackPanel>
 
     </Grid>
-</UserControl>
+</Window>
