+++ LibraryProject/LibraryApp/View/RegisterUser.xaml.cs
@@ -1,18 +1,27 @@
 ﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
 using System.Windows;
 using System.Windows.Controls;
+using System.Windows.Data;
+using System.Windows.Documents;
 using System.Windows.Input;
+using System.Windows.Media;
+using System.Windows.Media.Imaging;
+using System.Windows.Shapes;
 using LibraryApp.LibraryServiceReference;
 
-namespace LibraryApp
+namespace LibraryApp.View
 {
     /// <summary>
-    /// Interaction logic for Register.xaml
+    /// Interaction logic for RegisterUser.xaml
     /// </summary>
-    public partial class Register : UserControl
+    public partial class RegisterUser : Window
     {
         private ServiceClient _serviceClient = new ServiceClient();
-        public Register()
+        public RegisterUser()
         {
             InitializeComponent();
         }
@@ -20,7 +29,7 @@ namespace LibraryApp
         private void BtnRegister_OnClick(object sender, RoutedEventArgs e)
         {
             var result = _serviceClient.Register(TxtFirstname.Text, TxtLastname.Text, TxtAddress.Text, TxtPhone.Text,
-                   TxtUsername.Text, TxtPassword.Password);
+                TxtUsername.Text, TxtPassword.Password);
             MessageBox.Show(result ? "success" : "failed");
         }
 
@@ -31,8 +40,8 @@ namespace LibraryApp
         }
 
         private void UIElement_OnMouseDown(object sender, MouseButtonEventArgs e)
-        { 
-            
+        {
+            this.DragMove();
         }
     }
 }
